Time: 0 ms (100%), Space: 71.6 MB (80.41%) - LeetHub

+++ 0053-maximum-subarray/0053-maximum-subarray.cpp
@@ -1,20 +1,8 @@
 class Solution {
 public:
     int maxSubArray(vector<int>& nums) {
-        // long long maxi=INT_MIN;
-        // for(int i=0;i<nums.size();i++){
-        //     long long sum=0;
-        //     for(int j=i;j<nums.size();j++){
-        //         sum += nums[j];
-        //         maxi=max(sum,maxi);
-        //     }
-        // }
-        // return maxi;
-
-        // kadane's Algorithm
-
-        long long maxi=INT_MIN;
-        long long sum=0;
+        int maxi=INT_MIN;
+        int sum=0;
         for(int i=0;i<nums.size();i++){
             sum+=nums[i];
             if(sum>maxi){
